+++ src/TestUtils/Internal/TestAuthenticator.php
@@ -22,7 +22,6 @@ use Symfony\Component\Security\Http\Authenticator\Passport\SelfValidatingPasspor
 class TestAuthenticator extends AbstractAuthenticator
 {
     public const TEST_TOKEN = '42';
-    public const TEST_AUTHORIZATION_HEADER = 'Bearer '.self::TEST_TOKEN;
 
     private ?TestUser $user = null;